Treating CumQty as a double value

+++ quickfixj-examples/banzai/src/main/java/quickfix/examples/banzai/Order.java
@@ -26,7 +26,7 @@ public class Order implements Cloneable {
     private String symbol = null;
     private int quantity = 0;
     private int open = 0;
-    private int executed = 0;
+    private double executed = 0;
     private OrderSide side = OrderSide.BUY;
     private OrderType type = OrderType.MARKET;
     private OrderTIF tif = OrderTIF.DAY;
@@ -95,13 +95,11 @@ public class Order implements Cloneable {
         this.open = open;
     }
 
-    public int getExecuted() {
+    public double getExecuted() {
         return executed;
     }
 
-    public void setExecuted(int executed) {
-        this.executed = executed;
-    }
+    public void setExecuted(double executed) { this.executed = executed; }
 
     public OrderSide getSide() {
         return side;
